+++ src/AliceInventory/Controllers/AliceRequest.cs
@@ -1,18 +1,18 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace AliceInventory.Models
+namespace AliceInventory.Controllers
 {
     public class AliceRequest
     {
         [JsonProperty("meta")]
-        public MetaModel Meta { get; set; }
+        public Meta Meta { get; set; }
 
         [JsonProperty("request")]
-        public RequestModel Request { get; set; }
+        public Request Request { get; set; }
 
         [JsonProperty("session")]
-        public SessionModel Session { get; set; }
+        public Session Session { get; set; }
 
         [JsonProperty("version")]
         public string Version { get; set; }
